+++ web/src/features/interview-report/server/loaders/get-public-reports.ts
@@ -1,7 +1,7 @@
 import { createAdminClient } from "@mirai-gikai/supabase";
 import { unstable_cache } from "next/cache";
 import { CACHE_TAGS } from "@/lib/cache-tags";
-import type { PublicInterviewReport } from "../types";
+import type { PublicInterviewReport } from "../../shared/types";
 
 export const REPORTS_PER_PAGE = 20;
